requirements.txt
torch
torchvision
torchaudio
transformers
scikit-learn
pandas
numpy
pillow
streamlit
librosa
soundfile
